refactor(Router, App): App에서 라우팅 분리 #5
- 라우트 컴포넌트로 관심사를 분리
- App에서 문서화 중 todo 1개 삭제
